Generalize bm_chttp2_transport for other transports
Generalize bm_chttp2_transport.cc so that binder transport can be benchmarked
using the same framework. This is done by extracting common benchmark
logics out to a separate file bm_transport.h and make all functions
generic over a Fixture. We can then implement a Fixture class for binder
to make binder transport benchmark works (in upcoming PR).
